+++ actions/network_delete_globo.py
@@ -1,0 +1,14 @@
+#!/usr/bin/env python
+
+from lib.cloudstack import CloudStackAPI
+
+__all__ = [
+    'NetworkDeleteGlobo'
+]
+
+
+class NetworkDeleteGlobo(CloudStackAPI):
+    def run(self, url, apikey, secretkey, network_id):
+        cs = self.get_client(url, apikey, secretkey)
+
+        return cs.addNetworkViaGloboNetwork(id=network_id)